Encode to bytes when a string is passed
Hi, I was doing forensic on a Ubuntu 16.04 image, and the command "linux_bash" was outputting the following error
 `File "volatility/volatility/utils.py", line 164, in iterfind
    offset = data.find(string, 0)
TypeError: argument should be integer or bytes-like object, not 'str'`

I found that the "string" parameter wasn't always a bytes-like object. I added a try-except to encode it when needed. This is a really quick fix, and I don't know the code at all. There is maybe a proper alternative.

The command was working properly after this.